made cleaning up optional

+++ aggregators/PBSResultAggregatorWrapper.py
@@ -34,12 +34,15 @@ from tools.FileSystem import FileSystem
 
 
 class PBSResultAggregatorWrapper(JobResultAggregator):
-    def __init__(self, wrapped_aggregator, filename):
+    def __init__(self, wrapped_aggregator, filename, clean_up = False):
         self.wrapped_aggregator = wrapped_aggregator
         self.filename = filename
         
         # to keep track of all submitted results
         self.result_counter = 0
+        
+        # whether to delete job output
+        self.clean_up = clean_up
     
     def submit_result(self, result):
         # NOTE: this happens on the PBS
@@ -71,4 +74,5 @@ class PBSResultAggregatorWrapper(JobResultAggregator):
         return self.wrapped_aggregator.get_final_result()
 
     def clean_up(self):
-        FileSystem.delete_dir_failsafe(os.sep.join(self.filename.split(os.sep)[:-1]))
+        if self.clean_up:
+            FileSystem.delete_dir_failsafe(os.sep.join(self.filename.split(os.sep)[:-1]))
